refactor: rename gt context providers
