+++ packages/react/src/context.types.ts
@@ -1,9 +1,9 @@
-import { CSRGTProvider } from './provider/CSRGTProvider';
+import { BrowserGTProvider } from './provider/BrowserGTProvider';
 
 /**
  * Wrap GTProvider around the content that you want to translate
  */
-export const GTProvider: typeof CSRGTProvider = () => {
+export const GTProvider: typeof BrowserGTProvider = () => {
   throw new Error(
     'gt-react: You have imported a function from the dedicated types entrypoint. If you are seeing this error, it means something has gone wrong.'
   );
